Minor modifications to rebalanace tool (#4)
Handling a special case where the number of replicas for a partition could be greater than the replica count for the topic. In this case, the partition should not considered to be balanced even if there exists a subset of the replicas such that the replicas are in separate UDs and FDs and len(subset) = #replicaCount for topic.